PLIMS 3.2

+++ PLIMS/www/modules/PLIMS/update.sh
@@ -5,26 +5,26 @@ mos=`cat /tmp/plims_mos`
 fix_rights()
 {
   chmod 755 $mos/PLIMS/bin/rtmpdump
-  chmod 755 $mos/PLIMS/scripts/xLiveCZ/msdl
-  chmod 755 $mos/PLIMS/scripts/xLiveCZ/msdl.sh
+  chmod 755 $mos/PLIMS/scripts/xLiveCZ/msdl*
   chmod 755 $mos/PLIMS/scripts/xLiveCZ/nova.sh
-  rm -R $mos/PLIMS/scripts/xLiveCZ/rtmpdump
-  cd /usr/local/etc/mos/www/modules/PLIMS/scripts/xLiveCZ/
-  ln -s ./../../bin/rtmpdump rtmpdump
+  if [ ! -h $mos/PLIMS/scripts/xLiveCZ/rtmpdump ]; then
+      rm -R $mos/PLIMS/scripts/xLiveCZ/rtmpdump
+      cd $mos/PLIMS/scripts/xLiveCZ/
+      ln -s ./../../bin/rtmpdump rtmpdump
+  fi
 }
 
 create_backup()
 {
   echo "Backup ustawień przed instalacją"
-  cp -R  $mos/PLIMS/ustawienia.xml /tmp/ustawienia.xml   /tmp/
-  cp -R $mos/PLIMS/scripts/xLiveCZ/category/tv/WeebTV.xml  /tmp/
+  cp -R  $mos/PLIMS/ustawienia.xml  /tmp/
   
 }
 restore()
 {
  echo "Przywracanie ustawień"
  cp -R /tmp/ustawienia.xml $mos/PLIMS/
- cp -R /tmp/WeebTV.xml $mos/PLIMS/scripts/xLiveCZ/category/tv/
+ rm -R /tmp/ustawienia.xml
  
 }
 download()
@@ -44,15 +44,15 @@ check()
 
 if [ $mos = "/tmp/www/modules" ];
 then
-	create_backup;  
+#	create_backup;  
 	download;
 	check;
 	echo "Aktualizacja pakietu mos..."
 	pm update PLIMS now
-   fix_rights;
-   restore;
+	fix_rights;
+#	restore;
 else
-	create_backup;
+#	create_backup;
 	download;
 	check;
 	echo "Tworzenie tymczasowego katalogu.."
@@ -63,8 +63,8 @@ else
 	cp -R $mos/PLIMS_TEMP/www/modules/PLIMS/* $mos/PLIMS/
 	echo "Usuwanie tymczasowego katalogu.."
 	rm -R $mos/PLIMS_TEMP
-   fix_rights;
-   restore;
+	fix_rights;
+#	restore;
 fi
 
 echo -e "\nInstalacja przebiegła pomyślnie.\nZrestartuj PLIMS (wyjdź i wejdź) aby zobaczyć zmiany."
